Allow going mahjong by pressing mahjong button, and autofilling
Also fixes an infinite loop crash in bots when the last tile needed to be a sequence, but could be a pong - ex, a pair of 5s, and 567.

The bot failed to handle it, and looped. This change both fixes the issue, by using mahjong autocompletion, and blocks bot looping in most, if not all, cases.